requirements.txt
numpy==2.3.2
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
tzdata==2025.2
